Refactor hemisphere handler
